+++ electron/data/data.ts
@@ -4,10 +4,10 @@
 // this is where the data process ipc events are defined and handled
 //
 
-import { ipcMain, IpcMainInvokeEvent } from "electron";
+import { ipcMain, IpcMainInvokeEvent, webContents } from "electron";
 import { IStore, IStoreKeys, StoreKeys } from "~/Shared/store";
 import { win as mainWindow } from "../main/main";
-import { wheelWindow} from "../main/wheelOfNames"
+import { wheelWindow } from "../main/wheelOfNames";
 import { EChannels } from "~/Shared/channels";
 
 // load main ipc actions
@@ -20,39 +20,52 @@ let removeQue: string[] = [];
 
 // TODO claimed here IPC channel logic
 
-
 const forceUpdate = () => {
   wheelWindow?.webContents.send(EChannels.setDefaults);
   wheelWindow?.webContents.send(EChannels.reload);
-}
+};
 
 ipcMain.handle("forceUpdate", forceUpdate);
 
-
 const broadcastUpdate = <K extends IStoreKeys>(name: K, data: IStore[K]) => {
-  
   mainWindow?.webContents.send(EChannels.storeUpdate, name, data);
 };
 
 export const setStore = <K extends IStoreKeys>(name: K, data: IStore[K]) => {
   console.log("setting store", name, data);
-  
+
   store.set(name, data);
   broadcastUpdate(name, data);
 };
 
-export const handleAddWheelUser = (_: IpcMainInvokeEvent, entry: Entry, override=false) => {
-  console.log("addWheelUser", entry, override);
+const checkForDuplicates = (entry: Entry) => {
+  let data = store.get(StoreKeys.data, []);
+  return data.some((existingEntry) => {
+    if (entry.channelId && existingEntry.channelId) {
+      return existingEntry.channelId === entry.channelId;
+    } else {
+      return existingEntry.text === entry.text;
+    }
+  });
+};
+
+export const handleAddWheelUser = (
+  _: IpcMainInvokeEvent,
+  entry: Entry,
+  override = false
+) => {
+  entry.timestamp = Date.now();
   if (pause) {
     addQueue.push(entry);
     return true;
   }
-
+  entry.message = entry.channelId;
   let data = store.get(StoreKeys.data, []);
-  if (data.some((existingEntry) => existingEntry.text === entry.text)) {
+
+  if (checkForDuplicates(entry)) {
     console.log("Duplicate entry detected:", entry.text, override);
     // override the existing entry
-    if(override) {
+    if (override) {
       data = data.map((existingEntry) =>
         existingEntry.text === entry.text ? entry : existingEntry
       );
@@ -60,6 +73,8 @@ export const handleAddWheelUser = (_: IpcMainInvokeEvent, entry: Entry, override
       forceUpdate();
       return true;
     }
+    if (entry.channelId) handleUpdateActivity(_, entry.text, entry.channelId);
+
     return false;
   }
 
@@ -70,50 +85,94 @@ export const handleAddWheelUser = (_: IpcMainInvokeEvent, entry: Entry, override
 };
 
 export const handleRemoveWheelUser = (_: IpcMainInvokeEvent, name: string) => {
-    console.log("removeWheelUser", name);
-    if (pause) {
-        removeQue.push(name);
-        return true;
-    }
-    let data = store.get(StoreKeys.data);
-    const newdata = data.filter((entry) => entry.text !== name);
-
-    if (newdata.length === data.length) return false;
-
-    setStore(StoreKeys.data, newdata);
-    forceUpdate()
+  console.log("removeWheelUser", name);
+  if (pause) {
+    removeQue.push(name);
     return true;
+  }
+  let data = store.get(StoreKeys.data);
+  const newdata = data.filter((entry) => entry.text !== name);
+
+  if (newdata.length === data.length) return false;
+
+  setStore(StoreKeys.data, newdata);
+  forceUpdate();
+  return true;
 };
 
 export const handleUpdateWheelUser = (_: IpcMainInvokeEvent, entry: Entry) => {
   console.log("updateWheelUser", entry);
   let data = store.get(StoreKeys.data);
+  entry.timestamp = Date.now();
   data = data.map((existingEntry) =>
     existingEntry.text === entry.text ? entry : existingEntry
   );
   setStore(StoreKeys.data, data);
   forceUpdate();
-}
+};
 
 const handleResetClaims = () => {
-
   let data = store.get(StoreKeys.data);
   data = data.map((entry) => ({ ...entry, claimedHere: false } as Entry));
   setStore(StoreKeys.data, data);
   forceUpdate();
-}
+};
 
-ipcMain.handle('resetClaims', handleResetClaims);
+export const handleUpdateActivity = async (
+  _: IpcMainInvokeEvent,
+  displayName: string,
+  channelId: string
+) => {
+  
+  let entries = store.get(StoreKeys.data);
+
+  entries = entries.map((entry: Entry) => {
+    if (entry.channelId === channelId) {
+      return { ...entry, timestamp: Date.now(), message: channelId };
+    } else if (entry.text === displayName) {
+      return {
+        ...entry,
+        channelId,
+        timestamp: Date.now(),
+        message: channelId,
+      };
+    } else {
+      return entry;
+    }
+  });
+
+  if(wheelWindow){
+    console.log("WHEELOPEN", displayName, channelId);
+    
+  const lastconfig = JSON.parse(
+    await wheelWindow?.webContents.executeJavaScript(
+      `localStorage.getItem('LastWheelConfig')`
+    )
+  );
+
+  lastconfig.entries = entries;
+
+  const json = JSON.stringify(lastconfig);
+
+  await wheelWindow?.webContents.executeJavaScript(
+    `localStorage.setItem('LastWheelConfig', '${json}' )`
+  );
+}
+  setStore(StoreKeys.data, entries);
+};
+
+ipcMain.handle("resetClaims", handleResetClaims);
 ipcMain.handle("updateWheelUser", handleUpdateWheelUser);
 ipcMain.handle("addWheelUser", handleAddWheelUser);
 ipcMain.handle("removeWheelUser", handleRemoveWheelUser);
+ipcMain.handle("updateActivity", handleUpdateActivity);
 
 ipcMain.handle(
   "getStore",
   <K extends IStoreKeys>(_: IpcMainInvokeEvent, name: K): IStore[K] => {
     const data = store.get(name);
-    
-    return data
+
+    return data;
   }
 );
 
@@ -127,43 +186,43 @@ ipcMain.handle(
 ipcMain.handle("setPause", async (_: IpcMainInvokeEvent, value: boolean) => {
   console.log("setPause", value);
   saveConfig();
-  
+
   pause = value;
   if (!pause) {
-    
     await syncWithWheel();
     // create dummy ipc event to trigger the queue
     const event: IpcMainInvokeEvent = {} as IpcMainInvokeEvent;
     while (addQueue.length) {
-        const entry = addQueue.pop();
-        if (entry) handleAddWheelUser(event, entry);
+      const entry = addQueue.pop();
+      if (entry) handleAddWheelUser(event, entry);
     }
     while (removeQue.length) {
-        const name = removeQue.pop();
+      const name = removeQue.pop();
       if (name) handleRemoveWheelUser(event, name);
     }
-    // broadcastUpdate(StoreKeys.data, store.get(StoreKeys.data));
   }
 });
 
 const saveConfig = async () => {
   console.log("saveConfig");
   const response = await wheelWindow?.webContents.executeJavaScript(
-      `localStorage.getItem('LastWheelConfig')`
-    )
-    
-  const lastconfig = JSON.parse(response)
-    
+    `localStorage.getItem('LastWheelConfig')`
+  );
+
+  const lastconfig = JSON.parse(response);
+
   setStore(StoreKeys.lastconfig, lastconfig);
-}
+};
 
 ipcMain.handle("saveConfig", saveConfig);
 
 const syncWithWheel = async () => {
   console.log("syncWithWheel");
-  const lastconfig = JSON.parse(await wheelWindow?.webContents.executeJavaScript(
-    `localStorage.getItem('LastWheelConfig')`
-  ));
+  const lastconfig = JSON.parse(
+    await wheelWindow?.webContents.executeJavaScript(
+      `localStorage.getItem('LastWheelConfig')`
+    )
+  );
 
   // key:values in the lastconfig.entries may be missing
   // map though and set defaults if missing
@@ -177,8 +236,8 @@ const syncWithWheel = async () => {
     } as Entry;
   });
 
-  if(!lastconfig.isAdvanced) {
-    const duplicateObjects: {[key: string]: Entry} = {};
+  if (!lastconfig.isAdvanced) {
+    const duplicateObjects: { [key: string]: Entry } = {};
 
     entries.forEach((entry: Entry) => {
       if (duplicateObjects[entry.text]) {
@@ -190,14 +249,13 @@ const syncWithWheel = async () => {
 
     entries = Object.values(duplicateObjects);
     console.log("entries", entries);
-    
+
     setStore(StoreKeys.data, entries);
     forceUpdate();
     return;
   }
-  
-  setStore(StoreKeys.data, entries);
 
+  setStore(StoreKeys.data, entries);
 };
 
 ipcMain.handle("syncWithWheel", syncWithWheel);